// src/api/types/Project.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Project = {
  /**
   * @type string | undefined, uuid
   */
  id?: string
  /**
   * @type string | undefined
   */
  name?: string
  /**
   * @type string | undefined
   */
  slug?: string
  /**
   * @type string | undefined
   */
  shortDescription?: string
  /**
   * @type string | undefined
   */
  description?: string
  /**
   * @type array | undefined
   */
  skills?: string[]
  /**
   * @type string | undefined, uuid
   */
  linksId?: string
  /**
   * @type string | undefined, uuid
   */
  companyId?: string
  /**
   * @type string | undefined, uuid
   */
  coverId?: string
  /**
   * @type string | undefined, uuid
   */
  photoId?: string
  /**
   * @type string | undefined, date-time
   */
  createdAt?: string
  /**
   * @type string | undefined, date-time
   */
  updatedAt?: string
}