// src/api/types/Link.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Link = {
  /**
   * @type string | undefined, uuid
   */
  id?: string
  /**
   * @type string | undefined, uri
   */
  github?: string
  /**
   * @type string | undefined, uri
   */
  figma?: string
  /**
   * @type string | undefined, uri
   */
  publicUrl?: string
  /**
   * @type string | undefined, date-time
   */
  createdAt?: string
  /**
   * @type string | undefined, date-time
   */
  updatedAt?: string
}