// src/api/types/Company.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type Company = {
  /**
   * @type string | undefined, uuid
   */
  id?: string
  /**
   * @type string | undefined
   */
  name?: string
  /**
   * @type string | undefined
   */
  description?: string
  /**
   * @type string | undefined, date-time
   */
  startDate?: string
  /**
   * @type string | undefined, date-time
   */
  endDate?: string
  /**
   * @type string | undefined
   */
  responsibleName?: string
  /**
   * @type string | undefined
   */
  responsibleRole?: string
  /**
   * @type string | undefined, uuid
   */
  photoId?: string
  /**
   * @type string | undefined, date-time
   */
  createdAt?: string
  /**
   * @type string | undefined, date-time
   */
  updatedAt?: string
}